python_be/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.38.0
pydantic==2.12.4
pydantic-settings==2.8.1
python-dotenv==1.0.0
python-multipart==0.0.20
sentence-transformers==5.1.2
chromadb==1.3.5
pymupdf==1.26.6
torch==2.9.1
numpy==2.3.5
openai>=1.0.0
python-docx==1.1.2
